Extract desk connector interactions
