+++ apps/ui/src/ui-desks/connectors/utils.ts
@@ -1,10 +1,11 @@
 import { RichTextData } from '@wordpress/rich-text';
+import { sortByIndex, type Editor, type TLShape, type TLShapeId } from 'tldraw';
 import {
 	canvasShapeToDeskWidget,
 	isDeskConnectorCanvasShape,
 } from '@/ui-desks/desk/tldraw-adapter';
-import type { DeskWidget } from '@/ui-desks/widgets/types';
-import type { Editor, TLShape, TLShapeId } from 'tldraw';
+import { getWidgetDropHandler } from '@/ui-desks/widget-actions/drop-handlers';
+import type { DeskWidget, WidgetDropHandler } from '@/ui-desks/widgets/types';
 
 export interface DeskWidgetConnectionTarget {
 	shapeId: TLShapeId;
@@ -46,6 +47,21 @@ export function getOutgoingWidgetConnections(
 	return targets;
 }
 
+export function getCurrentSelectedWidgetConnectionTargets( editor: Editor ) {
+	const selectedShapeIds = editor.getSelectedShapeIds();
+	if ( selectedShapeIds.length !== 1 ) {
+		return [];
+	}
+
+	const [ selectedShapeId ] = selectedShapeIds;
+	const selectedShape = editor.getShape( selectedShapeId );
+	if ( ! selectedShape || ! canvasShapeToDeskWidget( selectedShape ) ) {
+		return [];
+	}
+
+	return getOutgoingWidgetConnections( editor, selectedShapeId );
+}
+
 export function getSelectedDeskConnectorToolbarItem(
 	editor: Editor
 ): SelectedDeskConnectorToolbarItem | null {
@@ -75,6 +91,54 @@ export function getSelectedDeskConnectorToolbarItem(
 		sourceLabel: getDeskWidgetConnectionLabel( sourceWidget ),
 		targetLabel: getDeskWidgetConnectionLabel( targetWidget ),
 	};
+}
+
+export function getWidgetShapeAtPagePoint( editor: Editor, point: { x: number; y: number } ) {
+	return editor.getShapeAtPoint( point, {
+		hitInside: true,
+		renderingOnly: true,
+		margin: editor.options.hitTestMargin / editor.getZoomLevel(),
+	} ) as TLShape | undefined;
+}
+
+export function getConnectableShapeAtPagePoint(
+	editor: Editor,
+	point: { x: number; y: number },
+	sourceShapeId: TLShapeId
+) {
+	const shape = getWidgetShapeAtPagePoint( editor, point );
+
+	if ( ! shape || shape.id === sourceShapeId || isDeskConnectorCanvasShape( shape ) ) {
+		return null;
+	}
+
+	return canvasShapeToDeskWidget( shape ) ? shape : null;
+}
+
+export function getWidgetDropTargetAtPagePoint(
+	editor: Editor,
+	point: { x: number; y: number },
+	sourceShapeId: TLShapeId,
+	sourceWidget: DeskWidget
+): { shape: TLShape; widget: DeskWidget; handler: WidgetDropHandler } | null {
+	const target = editor
+		.getCurrentPageShapes()
+		.filter( ( shape ) => shape.id !== sourceShapeId && ! isDeskConnectorCanvasShape( shape ) )
+		.map( ( shape ) => {
+			const bounds = editor.getShapePageBounds( shape.id );
+			const widget = canvasShapeToDeskWidget( shape );
+			const handler = widget ? getWidgetDropHandler( sourceWidget, widget ) : null;
+			return bounds && widget && handler && isPointInBounds( point, bounds )
+				? { shape, widget, handler }
+				: null;
+		} )
+		.filter(
+			( item ): item is { shape: TLShape; widget: DeskWidget; handler: WidgetDropHandler } =>
+				item !== null
+		)
+		.sort( ( first, second ) => sortByIndex( second.shape, first.shape ) )[ 0 ];
+
+	return target ?? null;
 }
 
 export function getDeskConnectorEndpoints(
@@ -178,4 +242,16 @@ function getRichTextPlainText( value: string ) {
 	}
 
 	return RichTextData.fromHTMLString( value ).toPlainText();
+}
+
+function isPointInBounds(
+	point: { x: number; y: number },
+	bounds: { minX: number; minY: number; maxX: number; maxY: number }
+) {
+	return (
+		point.x >= bounds.minX &&
+		point.x <= bounds.maxX &&
+		point.y >= bounds.minY &&
+		point.y <= bounds.maxY
+	);
 }
