+++ apps/ui/src/ui-desks/desk/provider/index.tsx
@@ -5,8 +5,6 @@ import {
 	getIndexAbove,
 	sortByIndex,
 	type Editor,
-	type TLArrowShape,
-	type TLEventInfo,
 	type TLShape,
 	type TLShapeId,
 	type TLShapePartial,
@@ -14,19 +12,19 @@ import {
 import { useConnector } from '@/data/core';
 import { useSites } from '@/data/queries/use-sites';
 import {
-	focusOnDeskShape,
+	focusConnectedWidgetInEditor,
+	removeSelectedConnectorFromEditor,
+	startConnectingWidgetInEditor,
+} from '@/ui-desks/connectors/editor-commands';
+import { useConnectorInteractions } from '@/ui-desks/connectors/use-connector-interactions';
+import {
+	getCurrentSelectedWidgetConnectionTargets,
 	getOutgoingWidgetConnections,
 	getSelectedDeskConnectorToolbarItem,
-} from '@/ui-desks/desk/connections';
+} from '@/ui-desks/connectors/utils';
 import {
-	CONNECTOR_COLOR,
-	CONNECTOR_DASH,
-	CONNECTOR_DEFAULT_BEND,
-	CONNECTOR_SHAPE_ID_PREFIX,
-	canvasShapeToDeskWidget,
 	getTemporaryDeskCanvasRecordMeta,
 	deskWidgetToCanvasShape,
-	isDeskConnectorCanvasShape,
 } from '@/ui-desks/desk/tldraw-adapter';
 import {
 	RECTANGLE_WIDGET_SHAPE_TYPE,
@@ -35,7 +33,6 @@ import {
 import { useStackInteractions } from '@/ui-desks/stacks/use-stack-interactions';
 import { useStackPressAnimation } from '@/ui-desks/stacks/use-stack-press-animation';
 import { createDeskWidget } from '@/ui-desks/widget-actions/create-widget';
-import { getWidgetDropHandler } from '@/ui-desks/widget-actions/drop-handlers';
 import { getWidgetEditAction } from '@/ui-desks/widget-actions/edit-action';
 import { getWidgetFileHandler } from '@/ui-desks/widget-actions/file-handlers';
 import {
@@ -70,8 +67,6 @@ import {
 import { useDeskPersistence } from './persistence';
 import { useDeskWidgetResolvers } from './resolvers';
 import type {
-	DeskWidget,
-	WidgetDropHandler,
 	WidgetHandlerLoading,
 	WidgetHandlerResult,
 	WidgetPastePayload,
@@ -147,6 +142,13 @@ export function DeskProvider( {
 	}, [ isRunningSite, selectedWidgetToolbarItem, siteId ] );
 
 	useStackInteractions( editor );
+	useConnectorInteractions( {
+		editor,
+		isHydrated,
+		isReadOnly,
+		pendingConnectorSourceId,
+		setPendingConnectorSourceId,
+	} );
 
 	useEffect( () => {
 		if ( deskConfigKeyRef.current === deskConfigKey ) {
@@ -193,26 +195,8 @@ export function DeskProvider( {
 				return previousShape;
 			}
 		);
-		const stopConnectorSelection = editor.store.listen(
-			() => {
-				const selectedShapeIds = editor.getSelectedShapeIds();
-				if ( selectedShapeIds.length === 0 ) {
-					return;
-				}
-
-				const nextSelectedShapeIds = selectedShapeIds.filter(
-					( shapeId ) => ! isDeskConnectorCanvasShape( editor.getShape( shapeId ) )
-				);
-				if ( nextSelectedShapeIds.length !== selectedShapeIds.length ) {
-					editor.setSelectedShapes( nextSelectedShapeIds );
-				}
-			},
-			{ scope: 'session' }
-		);
-
 		return () => {
 			stopShapeChanges();
-			stopConnectorSelection();
 		};
 	}, [ editor, isReadOnly ] );
 
@@ -396,277 +380,6 @@ export function DeskProvider( {
 				replaceTextShapeWithNote( editor, shape );
 			} );
 		} );
-	}, [ editor, isHydrated, isReadOnly ] );
-
-	useEffect( () => {
-		if ( ! editor ) {
-			return;
-		}
-
-		return editor.sideEffects.registerBeforeChangeHandler(
-			'shape',
-			( previousShape, nextShape ) => {
-				if ( ! isDeskConnectorCanvasShape( nextShape ) ) {
-					return nextShape;
-				}
-				if ( previousShape.x === nextShape.x && previousShape.y === nextShape.y ) {
-					return nextShape;
-				}
-				return {
-					...nextShape,
-					x: previousShape.x,
-					y: previousShape.y,
-				};
-			}
-		);
-	}, [ editor ] );
-
-	useEffect( () => {
-		if ( ! editor ) {
-			return;
-		}
-
-		return editor.sideEffects.registerAfterDeleteHandler( 'binding', ( binding ) => {
-			if ( binding.type !== 'arrow' ) {
-				return;
-			}
-
-			const connectorShape = editor.getShape( binding.fromId );
-			if ( ! isDeskConnectorCanvasShape( connectorShape ) ) {
-				return;
-			}
-
-			editor.deleteShape( connectorShape.id );
-		} );
-	}, [ editor ] );
-
-	useEffect( () => {
-		if ( ! editor || ! pendingConnectorSourceId ) {
-			return;
-		}
-
-		const sourceShape = editor.getShape( pendingConnectorSourceId );
-		if ( ! sourceShape || ! canvasShapeToDeskWidget( sourceShape ) ) {
-			setPendingConnectorSourceId( null );
-			return;
-		}
-
-		const sourceBounds = editor.getShapePageBounds( pendingConnectorSourceId );
-		const startPoint = toPlainPoint( sourceBounds?.center ?? editor.inputs.currentPagePoint );
-		const initialEndPoint = getInitialConnectorEndPoint(
-			startPoint,
-			toPlainPoint( editor.inputs.currentPagePoint )
-		);
-		const arrowId = createConnectorPreview(
-			editor,
-			pendingConnectorSourceId,
-			startPoint,
-			initialEndPoint
-		);
-
-		let completed = false;
-		const cancelConnection = () => {
-			setPendingConnectorSourceId( null );
-		};
-
-		const syncConnectorEnd = ( point: { x: number; y: number } ) => {
-			const hitShape = getConnectableShapeAtPagePoint( editor, point, pendingConnectorSourceId );
-			const hitBounds = hitShape ? editor.getShapePageBounds( hitShape.id ) : null;
-			updateConnectorEnd( editor, arrowId, toPlainPoint( hitBounds?.center ?? point ) );
-		};
-
-		const handleEvent = ( info: TLEventInfo ) => {
-			if ( info.type !== 'pointer' || info.name !== 'pointer_move' ) {
-				return;
-			}
-
-			syncConnectorEnd( toPlainPoint( editor.inputs.currentPagePoint ) );
-		};
-
-		const handlePointerMove = ( event: PointerEvent ) => {
-			syncConnectorEnd(
-				toPlainPoint( editor.screenToPage( { x: event.clientX, y: event.clientY } ) )
-			);
-		};
-
-		const handleClick = ( event: MouseEvent ) => {
-			const target = event.target as HTMLElement | null;
-			if ( target?.closest( '[data-ui-desks-context-menu]' ) ) {
-				return;
-			}
-
-			event.preventDefault();
-			event.stopPropagation();
-
-			const point = editor.screenToPage( { x: event.clientX, y: event.clientY } );
-			const targetShape = getConnectableShapeAtPagePoint( editor, point, pendingConnectorSourceId );
-			if ( ! targetShape ) {
-				cancelConnection();
-				return;
-			}
-
-			completed = true;
-			completeConnectorPreview( editor, arrowId, targetShape.id );
-			const targetBounds = editor.getShapePageBounds( targetShape.id );
-			if ( targetBounds ) {
-				updateConnectorEnd( editor, arrowId, toPlainPoint( targetBounds.center ) );
-			}
-			editor.setSelectedShapes( [ arrowId ] );
-			editor.focus();
-			setPendingConnectorSourceId( null );
-		};
-
-		const handleKeyDown = ( event: KeyboardEvent ) => {
-			if ( event.key === 'Escape' ) {
-				cancelConnection();
-			}
-		};
-
-		editor.on( 'event', handleEvent );
-		window.addEventListener( 'pointermove', handlePointerMove, true );
-		window.addEventListener( 'click', handleClick, true );
-		window.addEventListener( 'keydown', handleKeyDown );
-		return () => {
-			editor.off( 'event', handleEvent );
-			window.removeEventListener( 'pointermove', handlePointerMove, true );
-			window.removeEventListener( 'click', handleClick, true );
-			window.removeEventListener( 'keydown', handleKeyDown );
-			if ( ! completed && editor.getShape( arrowId ) ) {
-				editor.deleteShape( arrowId );
-			}
-		};
-	}, [ editor, pendingConnectorSourceId ] );
-
-	useEffect( () => {
-		if ( ! editor || ! isHydrated || isReadOnly ) {
-			return;
-		}
-
-		let source: {
-			shapeId: TLShapeId;
-			widget: DeskWidget;
-			type: string;
-			x: number;
-			y: number;
-		} | null = null;
-		let connectorPreviewId: TLArrowShape[ 'id' ] | null = null;
-		let activeTargetId: TLShapeId | null = null;
-		let didDrag = false;
-		let completed = false;
-
-		const removeConnectorPreview = () => {
-			if ( connectorPreviewId && editor.getShape( connectorPreviewId ) ) {
-				editor.deleteShape( connectorPreviewId );
-			}
-			connectorPreviewId = null;
-			activeTargetId = null;
-		};
-
-		const restoreSourcePosition = () => {
-			if ( ! source ) {
-				return;
-			}
-
-			const shape = editor.getShape( source.shapeId );
-			if ( ! shape || ( shape.x === source.x && shape.y === source.y ) ) {
-				return;
-			}
-
-			editor.updateShape( {
-				id: source.shapeId,
-				type: source.type as TLShape[ 'type' ],
-				x: source.x,
-				y: source.y,
-			} );
-		};
-
-		const cleanup = () => {
-			if ( ! completed ) {
-				removeConnectorPreview();
-			}
-			source = null;
-			connectorPreviewId = null;
-			activeTargetId = null;
-			didDrag = false;
-			completed = false;
-		};
-
-		const handleEvent = ( info: TLEventInfo ) => {
-			if ( info.type !== 'pointer' ) {
-				return;
-			}
-
-			if ( info.name === 'pointer_down' ) {
-				const shape = getWidgetShapeAtPagePoint( editor, editor.inputs.currentPagePoint );
-				const widget = shape ? canvasShapeToDeskWidget( shape ) : null;
-				source =
-					widget && shape
-						? { shapeId: shape.id, widget, type: shape.type, x: shape.x, y: shape.y }
-						: null;
-				connectorPreviewId = null;
-				activeTargetId = null;
-				didDrag = false;
-				completed = false;
-				return;
-			}
-
-			if ( ! source ) {
-				return;
-			}
-
-			if ( info.name === 'pointer_up' ) {
-				if ( didDrag && connectorPreviewId && activeTargetId ) {
-					completed = true;
-					completeConnectorPreview( editor, connectorPreviewId, activeTargetId );
-					const targetBounds = editor.getShapePageBounds( activeTargetId );
-					if ( targetBounds ) {
-						updateConnectorEnd( editor, connectorPreviewId, toPlainPoint( targetBounds.center ) );
-					}
-					restoreSourcePosition();
-				}
-				cleanup();
-				return;
-			}
-
-			if ( info.name !== 'pointer_move' || ! editor.inputs.isDragging ) {
-				return;
-			}
-
-			didDrag = true;
-			const point = toPlainPoint( editor.inputs.currentPagePoint );
-			const target = getWidgetDropTargetAtPagePoint( editor, point, source.shapeId, source.widget );
-			if ( ! target || target.handler.type !== 'connector' ) {
-				removeConnectorPreview();
-				return;
-			}
-
-			const targetBounds = editor.getShapePageBounds( target.shape.id );
-			if ( ! targetBounds ) {
-				removeConnectorPreview();
-				return;
-			}
-
-			if ( ! connectorPreviewId ) {
-				const sourceBounds = editor.getShapePageBounds( source.shapeId );
-				connectorPreviewId = createConnectorPreview(
-					editor,
-					source.shapeId,
-					toPlainPoint( sourceBounds?.center ?? point ),
-					toPlainPoint( targetBounds.center )
-				);
-			} else {
-				updateConnectorEnd( editor, connectorPreviewId, toPlainPoint( targetBounds.center ) );
-			}
-
-			activeTargetId = target.shape.id;
-			restoreSourcePosition();
-		};
-
-		editor.on( 'event', handleEvent );
-		return () => {
-			editor.off( 'event', handleEvent );
-			cleanup();
-		};
 	}, [ editor, isHydrated, isReadOnly ] );
 
 	const registerEditor = useCallback(
@@ -903,12 +616,10 @@ export function DeskProvider( {
 			return false;
 		}
 
-		const connector = getSelectedDeskConnectorToolbarItem( editor );
-		if ( ! connector ) {
+		if ( ! removeSelectedConnectorFromEditor( editor ) ) {
 			return false;
 		}
 
-		editor.deleteShape( connector.shapeId );
 		setSelectedConnectorToolbarItem( null );
 		return true;
 	}, [ editor, isHydrated, isReadOnly ] );
@@ -919,13 +630,11 @@ export function DeskProvider( {
 				return false;
 			}
 
-			const shape = editor.getShape( shapeId );
-			if ( ! shape || ! canvasShapeToDeskWidget( shape ) ) {
+			if ( ! startConnectingWidgetInEditor( editor, shapeId ) ) {
 				return false;
 			}
 
 			setPendingConnectorSourceId( shapeId );
-			editor.focus();
 			return true;
 		},
 		[ editor, isHydrated, isReadOnly ]
@@ -937,7 +646,7 @@ export function DeskProvider( {
 				return false;
 			}
 
-			return focusOnDeskShape( editor, shapeId );
+			return focusConnectedWidgetInEditor( editor, shapeId );
 		},
 		[ editor, isHydrated ]
 	);
@@ -1012,186 +721,6 @@ export function DeskProvider( {
 	} );
 
 	return <DeskContext.Provider value={ value }>{ children }</DeskContext.Provider>;
-}
-
-function createConnectorPreview(
-	editor: Editor,
-	sourceShapeId: TLShapeId,
-	startPoint: { x: number; y: number },
-	endPoint: { x: number; y: number }
-) {
-	const arrowId = createShapeId(
-		`${ CONNECTOR_SHAPE_ID_PREFIX }${ createWidgetId() }`
-	) as TLArrowShape[ 'id' ];
-	editor.createShape< TLArrowShape >( {
-		id: arrowId,
-		type: 'arrow',
-		meta: {
-			studioDeskConnector: true,
-		},
-		props: {
-			kind: 'arc',
-			color: CONNECTOR_COLOR,
-			dash: CONNECTOR_DASH,
-			size: 'm',
-			bend: CONNECTOR_DEFAULT_BEND,
-			arrowheadStart: 'dot',
-			arrowheadEnd: 'arrow',
-			start: startPoint,
-			end: endPoint,
-		},
-	} );
-	editor.createBindings( [
-		{
-			type: 'arrow',
-			fromId: arrowId,
-			toId: sourceShapeId,
-			props: {
-				terminal: 'start' as const,
-				normalizedAnchor: { x: 0.5, y: 0.5 },
-				isExact: false,
-				isPrecise: false,
-				snap: 'none' as const,
-			},
-		},
-	] );
-	return arrowId;
-}
-
-function completeConnectorPreview(
-	editor: Editor,
-	connectorShapeId: TLArrowShape[ 'id' ],
-	targetShapeId: TLShapeId
-) {
-	editor.createBindings( [
-		{
-			type: 'arrow',
-			fromId: connectorShapeId,
-			toId: targetShapeId,
-			props: {
-				terminal: 'end' as const,
-				normalizedAnchor: { x: 0.5, y: 0.5 },
-				isExact: false,
-				isPrecise: false,
-				snap: 'none' as const,
-			},
-		},
-	] );
-}
-
-function updateConnectorEnd(
-	editor: Editor,
-	connectorShapeId: TLArrowShape[ 'id' ],
-	endPoint: { x: number; y: number }
-) {
-	editor.updateShape< TLArrowShape >( {
-		id: connectorShapeId,
-		type: 'arrow',
-		props: {
-			end: endPoint,
-		},
-	} );
-}
-
-function getCurrentSelectedWidgetConnectionTargets( editor: Editor ) {
-	const selectedShapeIds = editor.getSelectedShapeIds();
-	if ( selectedShapeIds.length !== 1 ) {
-		return [];
-	}
-
-	const [ selectedShapeId ] = selectedShapeIds;
-	const selectedShape = editor.getShape( selectedShapeId );
-	if ( ! selectedShape || ! canvasShapeToDeskWidget( selectedShape ) ) {
-		return [];
-	}
-
-	return getOutgoingWidgetConnections( editor, selectedShapeId );
-}
-
-function getWidgetShapeAtPagePoint( editor: Editor, point: { x: number; y: number } ) {
-	return editor.getShapeAtPoint( point, {
-		hitInside: true,
-		renderingOnly: true,
-		margin: editor.options.hitTestMargin / editor.getZoomLevel(),
-	} ) as TLShape | undefined;
-}
-
-function getWidgetDropTargetAtPagePoint(
-	editor: Editor,
-	point: { x: number; y: number },
-	sourceShapeId: TLShapeId,
-	sourceWidget: DeskWidget
-): { shape: TLShape; widget: DeskWidget; handler: WidgetDropHandler } | null {
-	const target = editor
-		.getCurrentPageShapes()
-		.filter( ( shape ) => shape.id !== sourceShapeId && ! isDeskConnectorCanvasShape( shape ) )
-		.map( ( shape ) => {
-			const bounds = editor.getShapePageBounds( shape.id );
-			const widget = canvasShapeToDeskWidget( shape );
-			const handler = widget ? getWidgetDropHandler( sourceWidget, widget ) : null;
-			return bounds && widget && handler && isPointInBounds( point, bounds )
-				? { shape, widget, handler }
-				: null;
-		} )
-		.filter(
-			( item ): item is { shape: TLShape; widget: DeskWidget; handler: WidgetDropHandler } =>
-				item !== null
-		)
-		.sort( ( first, second ) => sortByIndex( second.shape, first.shape ) )[ 0 ];
-
-	return target ?? null;
-}
-
-function getConnectableShapeAtPagePoint(
-	editor: Editor,
-	point: { x: number; y: number },
-	sourceShapeId: TLShapeId
-) {
-	const shape = editor.getShapeAtPoint( point, {
-		hitInside: true,
-		renderingOnly: true,
-		margin: editor.options.hitTestMargin / editor.getZoomLevel(),
-	} ) as TLShape | undefined;
-
-	if ( ! shape || shape.id === sourceShapeId || isDeskConnectorCanvasShape( shape ) ) {
-		return null;
-	}
-
-	return canvasShapeToDeskWidget( shape ) ? shape : null;
-}
-
-function isPointInBounds(
-	point: { x: number; y: number },
-	bounds: { minX: number; minY: number; maxX: number; maxY: number }
-) {
-	return (
-		point.x >= bounds.minX &&
-		point.x <= bounds.maxX &&
-		point.y >= bounds.minY &&
-		point.y <= bounds.maxY
-	);
-}
-
-function toPlainPoint( point: { x: number; y: number } ) {
-	return {
-		x: point.x,
-		y: point.y,
-	};
-}
-
-function getInitialConnectorEndPoint(
-	startPoint: { x: number; y: number },
-	cursorPoint: { x: number; y: number }
-) {
-	const distance = Math.hypot( cursorPoint.x - startPoint.x, cursorPoint.y - startPoint.y );
-	if ( distance >= 8 ) {
-		return cursorPoint;
-	}
-
-	return {
-		x: startPoint.x + 96,
-		y: startPoint.y,
-	};
 }
 
 function replaceTextShapeWithNote( editor: Editor, shape: TLShape ) {
